Implement the time package
Also add an internal package, add index.js for each package so all you
have to do is import that. Support loading tzinfo files from specified
locations on disk.

Also add sync.Once which is a little silly when JS is single threaded,
but at least the implementation was easy.

Add a bytes.Slice type so we can have a Buffer that can actually grow.
I'm not sure about the API's for those, quite yet.